Integration

+++ lib/ui/menu/viewmodel/modifier_viewmodel.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/foundation.dart' show ChangeNotifier;
 import 'package:street_cart_pos/data/repositories/menu_repository.dart';
 import 'package:street_cart_pos/domain/models/product_model.dart';
-import 'package:uuid/uuid.dart';
 
 class ModifierViewModel extends ChangeNotifier {
   final MenuRepository _repository = MenuRepository();
+  final Set<String> _pendingDeleteIds = {};
+  bool _disposed = false;
 
   ModifierViewModel() {
     _repository.addListener(notifyListeners);
@@ -12,46 +13,36 @@ class ModifierViewModel extends ChangeNotifier {
 
   @override
   void dispose() {
+    _disposed = true;
     _repository.removeListener(notifyListeners);
     super.dispose();
   }
 
-  List<ModifierGroup> get modifierGroups => List.unmodifiable(_repository.modifierGroups);
+  List<ModifierGroup> get modifierGroups => List.unmodifiable(
+    _repository.modifierGroups.where((g) => !_pendingDeleteIds.contains(g.id)),
+  );
 
-  Future<void> addModifierGroup(String name, int optionCount) async {
-    // Create dummy options to match the count returned by the form
-    final options = List.generate(
-      optionCount,
-      (index) => ModifierOptions(
-        id: const Uuid().v4(),
-        name: 'Option ${index + 1}',
-      ),
-    );
-
-    await _repository.addModifierGroup(ModifierGroup(
-      id: const Uuid().v4(),
-      name: name,
-      modifierOptions: options,
-    ));
+  Future<void> addModifierGroup(ModifierGroup group) async {
+    await _repository.addModifierGroup(group);
   }
 
-  Future<void> updateModifierGroup(ModifierGroup group, String name, int optionCount) async {
-    final options = List.generate(
-      optionCount,
-      (index) => ModifierOptions(
-        id: const Uuid().v4(),
-        name: 'Option ${index + 1}',
-      ),
-    );
-
-    await _repository.updateModifierGroup(ModifierGroup(
-      id: group.id,
-      name: name,
-      modifierOptions: options,
-    ));
+  Future<void> updateModifierGroup(ModifierGroup group) async {
+    await _repository.updateModifierGroup(group);
   }
 
   Future<void> deleteModifierGroup(ModifierGroup group) async {
-    await _repository.deleteModifierGroup(group.id);
+    if (_pendingDeleteIds.contains(group.id)) return;
+    _pendingDeleteIds.add(group.id);
+    if (!_disposed) notifyListeners();
+
+    try {
+      await _repository.deleteModifierGroup(group.id);
+    } catch (_) {
+      _pendingDeleteIds.remove(group.id);
+      if (!_disposed) notifyListeners();
+      rethrow;
+    } finally {
+      _pendingDeleteIds.remove(group.id);
+    }
   }
 }
